add '_' to "hide" some utility functions in trees.py; move add_classifier_legend to classifiers.py
